move usePathInfo into fs/common/hooks
