+++ shared/fs/common/index.tsx
@@ -12,4 +12,11 @@ export {default as FolderViewFilterIcon} from './folder-view-filter-icon'
 export {default as SyncStatus} from './sync-status-container'
 export {default as Filename} from './filename'
 
-export {useFsPathMetadata, useFsChildren, useFsTlfs, useFsJournalStatus, useFsOnlineStatus} from './hooks'
+export {
+  useFsPathMetadata,
+  useFsChildren,
+  useFsTlfs,
+  useFsJournalStatus,
+  useFsOnlineStatus,
+  useFsPathInfo,
+} from './hooks'
