+++ shared/fs/common/path-info.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import * as Types from '../../constants/types/fs'
 import * as Constants from '../../constants/fs'
 import * as Container from '../../util/container'
-import * as FsGen from '../../actions/fs-gen'
+import {useFsPathInfo} from './hooks'
 import * as Styles from '../../styles'
 import * as Kb from '../../common-adapters'
 
@@ -10,22 +10,6 @@ type PathInfoProps = {
   containerStyle?: Styles.StylesCrossPlatform
   knownPathInfo?: Types.PathInfo
   path: Types.Path
-}
-
-const usePathInfo = (path: Types.Path, knownPathInfo: Types.PathInfo): Types.PathInfo => {
-  const pathInfo = Container.useSelector(state => state.fs.pathInfos.get(path, Constants.emptyPathInfo))
-  const dispatch = Container.useDispatch()
-  const alreadyKnown = knownPathInfo !== Constants.emptyPathInfo
-  React.useEffect(() => {
-    if (alreadyKnown) {
-      dispatch(FsGen.createLoadedPathInfo({path, pathInfo: knownPathInfo}))
-    } else if (pathInfo === Constants.emptyPathInfo) {
-      // We only need to load if it's empty. This never changes once we have
-      // it.
-      dispatch(FsGen.createLoadPathInfo({path}))
-    }
-  }, [path, alreadyKnown, knownPathInfo, pathInfo, dispatch])
-  return alreadyKnown ? knownPathInfo : pathInfo
 }
 
 const useMountPointPath = (platformAfterMountPath: string) => {
@@ -40,7 +24,7 @@ const useMountPointPath = (platformAfterMountPath: string) => {
 // name it different because of a bug in eslint that warns above in Props:
 //   11:25  error  'PathInfo' was used before it was defined  no-use-before-define
 const PathInfo_ = (props: PathInfoProps) => {
-  const pathInfo = usePathInfo(props.path, props.knownPathInfo || Constants.emptyPathInfo)
+  const pathInfo = useFsPathInfo(props.path, props.knownPathInfo || Constants.emptyPathInfo)
   const mountPointPath = useMountPointPath(pathInfo.platformAfterMountPath)
   return (
     <Kb.Box2 direction="vertical" style={props.containerStyle} fullWidth={true}>
